# refactor: speed up ColumnBuilder::repeat(Scalar::Null).