--- a/src/query/expression/src/types/binary.rs
+++ b/src/query/expression/src/types/binary.rs
@@ -346,6 +346,14 @@ impl BinaryColumnBuilder {
         }
     }
 
+    pub fn repeat_null(n: usize) -> Self {
+        BinaryColumnBuilder {
+            data: vec![],
+            offsets: vec![0; n + 1],
+            need_estimated: false,
+        }
+    }
+
     pub fn len(&self) -> usize {
         self.offsets.len() - 1
     }