--- a/src/query/expression/src/types/binary.rs
+++ b/src/query/expression/src/types/binary.rs
@@ -346,7 +346,7 @@ impl BinaryColumnBuilder {
         }
     }
 
-    pub fn repeat_null(n: usize) -> Self {
+    pub fn repeat_default(n: usize) -> Self {
         BinaryColumnBuilder {
             data: vec![],
             offsets: vec![0; n + 1],