I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

### Problem Description

The original implementation calls push_default n times, which is slow because each call is recursive and includes pattern matching.

### Impact

The repeat method is used when flattening Value::Scalar to Value::Column. This approach significantly affects or even dominates overall performance when there are many null columns.

related issue: https://github.com/datafuselabs/databend/issues/15908
https://github.com/datafuselabs/databend/pull/15880 will be faster after this

### Improvement

This change introduces an ColumnBuilder::repeat_null that create the whole builder at once;



### Performance Gains

bench with Criterion:
 create column of type `float32 null` with n rows.

```
null/old/100000  time:   [247.14 µs 248.32 µs 249.81 µs]
null/new/100000  time:   [31.723 µs 31.756 µs 31.791 µs]

null/old/10000000
                        time:   [29.333 ms 29.385 ms 29.447 ms]
null/new/10000000
                        time:   [5.3346 ms 5.4707 ms 5.6110 ms]

```

## Tests

- [ ] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):