@sundy-li 2024-07-01
can we add another method named `push_defaults(len: usize)`

```
 for _ in 0..n {
                        builder.push_default();
                    }
```

--->

```
builder.push_defaults(n);
```