@youngsofun 2024-07-01
> can we add another method named `push_defaults(len: usize)`
> 
> ```
> for _ in 0..n {
>    builder.push_default();
> }
> ```
> 
> --->
> 
> ```
> builder.push_defaults(n);
> ```


I have considered this approach  as well

1.  less straight forward and efficient when use in repeat
2.  no usage of push_defaults(n); with n > 1 found, except in repeat().

![image](https://github.com/datafuselabs/databend/assets/5782159/553f84f3-73e0-48a1-9259-1f2c476616e3)